go: package errors separately from types
