+++ sdk/go/core/docsgen/api/services.rst
@@ -105,7 +105,7 @@ Executor
 	:param topEntity: Provide an instance of :ref:`Entity <types-entity>` only when expecting data to be returned
 	:return: Any data the resulting from the operation when provided topEntity parameter
 	:rtype: :ref:`Entity <types-entity>` or ``nil``
-	:raises: :go:struct:`YError<ydk/types/YError>` If error has occurred
+	:raises: :go:struct:`YError<ydk/errors/YError>` If error has occurred
 
 	Possible Errors:
 	* a server side error
@@ -126,7 +126,7 @@ Netconf
 	:param persistID: An ``int``
 	:return: whether or not the operation succeeded
 	:rtype: ``bool``
-	:raises: :go:struct:`YError<ydk/types/YError>` If error has occurred
+	:raises: :go:struct:`YError<ydk/errors/YError>` If error has occurred
 
 .. function:: (ns *NetconfService) CloseSession(provider ServiceProvider)
 
@@ -135,7 +135,7 @@ Netconf
 	:param provider: An instance of :go:struct:`NetconfServiceProvider<ydk/providers/NetconfServiceProvider>`
 	:return: whether or not the operation succeeded
 	:rtype: ``bool``
-	:raises: :go:struct:`YError<ydk/types/YError>` If error has occurred
+	:raises: :go:struct:`YError<ydk/errors/YError>` If error has occurred
 
 .. function:: (ns *NetconfService) Commit(
 	provider ServiceProvider, confirmed bool, confirmTimeOut, persist, persistID int)
@@ -149,7 +149,7 @@ Netconf
 	:param persistID: An ``int`` that is given in order to commit a persistent confirmed commit
 	:return: whether or not the operation succeeded
 	:rtype: ``bool``
-	:raises: :go:struct:`YError<ydk/types/YError>` If error has occurred
+	:raises: :go:struct:`YError<ydk/errors/YError>` If error has occurred
 
 .. function:: (ns *NetconfService) CopyConfig(
 	provider ServiceProvider, target, sourceDS DataStore, sourceEntity Entity, url string)
@@ -164,7 +164,7 @@ Netconf
 	:param url: A ``string`` representing the configuration url
 	:return: whether or not the operation succeeded
 	:rtype: ``bool``
-	:raises: :go:struct:`YError<ydk/types/YError>` If error has occurred
+	:raises: :go:struct:`YError<ydk/errors/YError>` If error has occurred
 
 .. function:: (ns *NetconfService) DeleteConfig(provider ServiceProvider, target DataStore, url string)
 
@@ -175,7 +175,7 @@ Netconf
 	:param url: A ``string`` representing the configuration url
 	:return: whether or not the operation succeeded
 	:rtype: ``bool``
-	:raises: :go:struct:`YError<ydk/types/YError>` If error has occurred
+	:raises: :go:struct:`YError<ydk/errors/YError>` If error has occurred
 
 .. function:: (ns *NetconfService) DiscardChanges(provider ServiceProvider)
 
@@ -184,7 +184,7 @@ Netconf
 	:param provider: An instance of :go:struct:`NetconfServiceProvider<ydk/providers/NetconfServiceProvider>`
 	:return: whether or not the operation succeeded
 	:rtype: ``bool``
-	:raises: :go:struct:`YError<ydk/types/YError>` If error has occurred
+	:raises: :go:struct:`YError<ydk/errors/YError>` If error has occurred
 
 .. function:: (ns *NetconfService) EditConfig(
     provider ServiceProvider, target DataStore, config Entity, defaultOper, testOp, errorOp string)
@@ -199,7 +199,7 @@ Netconf
 	:param errOp: A ``string`` that can be set to ``stop-on-error``, ``continue-on-error``, or ``rollback-on-error``
 	:return: whether or not the operation succeeded
 	:rtype: ``bool``
-	:raises: :go:struct:`YError<ydk/types/YError>` If error has occurred
+	:raises: :go:struct:`YError<ydk/errors/YError>` If error has occurred
 
 .. function:: (ns *NetconfService) GetConfig(provider ServiceProvider, source DataStore, filter Entity)
 
@@ -210,7 +210,7 @@ Netconf
 	:param filter: An instance of :ref:`Entity <types-entity>` which specifies the portion of the system configuration and state data to retrieve
 	:return: The requested data
 	:rtype: :ref:`Entity <types-entity>`
-	:raises: :go:struct:`YError<ydk/types/YError>` If error has occurred
+	:raises: :go:struct:`YError<ydk/errors/YError>` If error has occurred
 
 
 .. function:: (ns *NetconfService) Get(provider ServiceProvider, filter Entity)
@@ -221,7 +221,7 @@ Netconf
 	:param filter: An instance of :ref:`Entity <types-entity>` which specifies the portion of the system configuration and state data to retrieve
 	:return: The requested data
 	:rtype: :ref:`Entity <types-entity>`
-	:raises: :go:struct:`YError<ydk/types/YError>` If error has occurred
+	:raises: :go:struct:`YError<ydk/errors/YError>` If error has occurred
 
 .. function:: (ns *NetconfService) KillSession(provider ServiceProvider, sessionID int)
 
@@ -231,7 +231,7 @@ Netconf
 	:param sessionID: An ``int`` that is the session identifier of the NETCONF session to be terminated
 	:return: whether or not the operation succeeded
 	:rtype: ``bool``
-	:raises: :go:struct:`YError<ydk/types/YError>` If error has occurred
+	:raises: :go:struct:`YError<ydk/errors/YError>` If error has occurred
 
 .. function:: (ns *NetconfService) Lock(provider ServiceProvider, target DataStore)
 
@@ -241,7 +241,7 @@ Netconf
 	:param target: An instance of :ref:`DataStore <datastore-ydk>` representing the configuration to lock
 	:return: whether or not the operation succeeded
 	:rtype: ``bool``
-	:raises: :go:struct:`YError<ydk/types/YError>` If error has occurred
+	:raises: :go:struct:`YError<ydk/errors/YError>` If error has occurred
 
 .. function:: (ns *NetconfService) Unlock(provider ServiceProvider, target DataStore)
 
@@ -251,7 +251,7 @@ Netconf
 	:param target: An instance of :ref:`DataStore <datastore-ydk>` representing the configuration to unlock
 	:return: whether or not the operation succeeded
 	:rtype: ``bool``
-	:raises: :go:struct:`YError<ydk/types/YError>` If error has occurred
+	:raises: :go:struct:`YError<ydk/errors/YError>` If error has occurred
 
 .. function:: (ns *NetconfService) Validate(
 	provider ServiceProvider, sourceDS DataStore, sourceEntity Entity, url string)
@@ -264,4 +264,4 @@ Netconf
 	:param url: A ``string`` representing the configuration url
 	:return: whether or not the operation succeeded
 	:rtype: ``bool``
-	:raises: :go:struct:`YError<ydk/types/YError>` If error has occurred
+	:raises: :go:struct:`YError<ydk/errors/YError>` If error has occurred
