+++ sdk/go/core/docsgen/api/types/_types_api.rst
@@ -11,4 +11,3 @@ Types
     base_constants/yfilters.rst
     base_constants/ytype.rst
     ydk_types.rst
-    errors.rst
