+++ sdk/go/core/docsgen/api/ydk_api.rst
@@ -5,6 +5,7 @@ YDK API
 .. toctree::
     :maxdepth: 3
 
-    services/services.rst
-    providers/providers.rst
+    services.rst
+    providers.rst
     types/_types_api.rst
+    errors.rst
